Split in Section Processors instead of a very large command
